cleaner-job/requirements.txt
certifi==2024.2.2
charset-normalizer==3.3.2
ecdsa==0.13.3
idna==3.7
jws==0.1.3
requests==2.31.0
urllib3==2.2.1

python-dotenv==1.0.1
azure-core==1.30.2
azure-storage-blob==12.20.0
